Added image transparency in GamePlayer, mouse blanks on engine execution.
In GamePlayer, a loaded image may now have a transparent color. By default,
the default color is magenta (FF00FF). If this color doesn't occur in the
image, there will be no transparency.

Also, a command as been added to blank the mouse cursor on the SDL screen
during execution.