Restricted GamePlayer movement in GameStaticMovementState, conformed style
In GameStaticMovementState, thePlayer is no longer allowed to move outside
the bounds of the screen.

We are now making an effort toward conforming to the style standards laid out
by the Google C++ Style Guide, which can be found at

	http://google-styleguide.googlecode.com/svn/trunk/cppguide.xml

Corrections to file formatting to meet these guidelines should be done
without prejudice.